+++ src/infrastructure/customer/mocks/customer.repository.mock.ts
@@ -1,5 +1,5 @@
-import Customer from "../../domain/customer/entity/customer";
-import CustomerRepositoryInterface from "../../domain/customer/repository/customer.repository.interface";
+import Customer from "../../../domain/customer/entity/customer";
+import CustomerRepositoryInterface from "../../../domain/customer/repository/customer.repository.interface";
 
 export default
   class CustomerRepositoryMock
