+++ src/infrastructure/customer/repository/sequelize/customer.repository.ts
@@ -1,7 +1,7 @@
-import Address from "../../domain/customer/value-object/address";
-import Customer from "../../domain/customer/entity/customer";
-import CustomerRepositoryInterface from "../../domain/customer/repository/customer.repository.interface";
-import CustomerModel from "../db/sequelize/model/customer.model";
+import Address from "../../../../domain/customer/value-object/address";
+import Customer from "../../../../domain/customer/entity/customer";
+import CustomerRepositoryInterface from "../../../../domain/customer/repository/customer.repository.interface";
+import CustomerModel from "./customer.model";
 
 export default
   class CustomerRepository
