+++ src/infrastructure/product/repository/sequelize/product.repository.ts
@@ -1,6 +1,6 @@
-import Product from "../../domain/product/entity/product";
-import ProductRepositoryInterface from "../../domain/product/repository/product.repository.interface";
-import ProductModel from "../db/sequelize/model/product.model";
+import Product from "../../../../domain/product/entity/product";
+import ProductRepositoryInterface from "../../../../domain/product/repository/product.repository.interface";
+import ProductModel from "./product.model";
 
 export default class ProductRepository implements ProductRepositoryInterface {
   async create(product: Product): Promise<void> {
